pyproject.toml
[project]
name = "scenenapse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "boltons>=25.0.0",
    "diffusers>=0.35.2",
    "dspy-ai>=2.5.0",
    "google-genai>=1.53.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "torchvision>=0.24.1",
    "transformers>=4.40.0",
    "ujson>=5.11.0",
]
